[Chore #107707438] Codebase clean up, moved login required mixin to appropriate view and removed erroneous route
