+++ codango/userprofile/urls.py
@@ -1,17 +1,12 @@
-from django.conf import settings
 from django.conf.urls import url
 from userprofile import views
 
 
 urlpatterns = [
-
     url(r'^(?P<username>\w+)$',
         views.UserProfileDetailView.as_view(), name='user_profile'),
     url(r'^(?P<username>\w+)/edit$',
         views.UserProfileEditView.as_view(), name='edit_user_profile'),
-    url(r'^media/(?P<path>.*)$', 'django.views.static.serve', {
-        'document_root': settings.MEDIA_ROOT,
-    }),
     url(r'^(?P<username>\w+)/follow$',
         views.FollowUserView.as_view(), name='follow_user'),
     url(r'^(?P<username>\w+)/(?P<direction>following|followers)',
